+++ packages/std/src/amp/messages.rs
@@ -434,7 +434,7 @@ impl AMPPkt {
     /// Generates a SubMsg to send the AMPPkt to the kernel
     pub fn to_sub_msg(
         &self,
-        address: impl Into<String>,
+        address: Addr,
         funds: Option<Vec<Coin>>,
         id: u64,
     ) -> Result<SubMsg, ContractError> {
@@ -695,7 +695,7 @@ mod tests {
 
         let pkt = AMPPkt::new(ORIGIN, PREVIOUS_SENDER, vec![msg.clone()]);
 
-        let sub_msg = pkt.to_sub_msg(KERNEL, None, 1).unwrap();
+        let sub_msg = pkt.to_sub_msg(Addr::unchecked(KERNEL), None, 1).unwrap();
 
         let expected_msg = ExecuteMsg::AMPReceive(AMPPkt::new(ORIGIN, PREVIOUS_SENDER, vec![msg]));
         assert_eq!(sub_msg.id, 1);